requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
django-filter==2.0.0
django-sendsms==0.3.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drfaddons==0.1.0
idna==2.7
pkginfo==1.4.2
PyJWT==1.6.4
pytz==2018.7
requests==2.20.0
requests-toolbelt==0.8.0
tqdm==4.28.1
urllib3==1.24.1
